Migrate to pyfakefs
