--- tests/files/test.c
@@ -1,9 +1,0 @@
-// TODO test
-
-#include <test.h>
-
-int main()
-{
-    printf("Hello World!"); //TODO test 2
-    return 0;
-}
