--- tests/files/test.cpp
@@ -1,9 +1,0 @@
-// TODO test
-
-#include <test.hpp>
-
-int main()
-{
-    std::cout << "Hello World!\n"; //TODO test 2
-    return 0;
-}
