--- tests/files/test.cs
@@ -1,10 +1,0 @@
-// TODO test
-namespace HelloWorld
-{
-    class Hello {         
-        static void Main(string[] args)
-        {
-            System.Console.WriteLine("Hello World!"); // TODO test 2
-        }
-    }
-}
