--- tests/files/test.go
@@ -1,8 +1,0 @@
-package main
-
-import "fmt"
-
-// TODO test
-func main() {
-    fmt.Println("hello world") // TODO test 2
-}
