--- tests/files/test.h
@@ -1,3 +1,0 @@
-//TODO test
-
-#include <stdio.h> //TODO test 2
