--- tests/files/test.hpp
@@ -1,3 +1,0 @@
-//TODO test
-
-#include <iostream> //TODO test 2
