--- tests/files/test.java
@@ -1,6 +1,0 @@
-// TODO test
-class HelloWorld {
-    public static void main(String[] args) {
-        System.out.println("Hello, World!"); // TODO test 2
-    }
-}
