--- tests/files/test.js
@@ -1,4 +1,0 @@
-// TODO test
-
-var foo = "Hello World!"
-console.log(foo) //TODO test 2
