--- tests/files/test.py
@@ -1,9 +1,0 @@
-# TODO test
-
-
-def main():
-    print("Hello World")  # TODO test 2
-
-
-if __name__ == '__main__':
-    main()
