--- tests/files/test.rb
@@ -1,2 +1,0 @@
-# TODO test
-puts 'Hello, world!' # TODO test 2
