--- tests/files/test.tex
@@ -1,5 +1,0 @@
-\begin{document}
-% TODO test
-Test % TODO test 2
-\end{document}
-
